+++ bf-asm/counter.cpp
@@ -126,13 +126,13 @@ int CounterTable::address_shift() const {
     return counter_shifts[format->groups()];
 }
 
-unsigned CounterTable::determine_shiftcount(Table::Call &call, int group, int word,
+unsigned CounterTable::determine_shiftcount(Table::Call &call, int group, unsigned word,
         int tcam_shift) const {
     if (call.args[0].name() && strcmp(call.args[0].name(), "$DIRECT") == 0) {
         return direct_shiftcount() + tcam_shift;
     } else if (call.args[0].field()) {
-        assert(call.args[0].field()->by_group[group]->bit(0)/128U == word);
-        return call.args[0].field()->by_group[group]->bit(0)%128U + indirect_shiftcount();
+        assert(int(call.args[0].field()->by_group[group]->bit(0)/128) == word);
+        return call.args[0].field()->by_group[group]->bit(0)%128 + indirect_shiftcount();
     } else if (call.args[1].field()) {
         return call.args[1].field()->bit(0) + STAT_ADDRESS_ZERO_PAD;
     }
